Refactor: Removed legacy TAK and made it into TAK-V2
