+++ docs/how_to_run/index.rst
@@ -62,7 +62,7 @@ Common Options
 
 For more details on arguments, see :ref:`module-utils`.
 
-For remote checkpoint loading (including Hugging Face URLs), TAKv2 Fisher cache loading from remote sources, and artifact upload to Hugging Face, see :doc:`getting_started/checkpoints`.
+For remote checkpoint loading (including Hugging Face URLs), TAK Fisher cache loading from remote sources, and artifact upload to Hugging Face, see :doc:`getting_started/checkpoints`.
 
 Examples
 ~~~~~~~~
